Imported Upstream version 9.1.6.0

+++ spec/ruby/core/process/fixtures/common.rb
@@ -1,35 +1,30 @@
 module ProcessSpecs
   class Daemonizer
-    attr_reader :input, :data, :signal
+    attr_reader :input, :data
 
     def initialize
+      # Fast feedback for implementations without Process.daemon
+      raise NotImplementedError, "Process.daemon is not implemented" unless Process.respond_to? :daemon
+
       @script = fixture __FILE__, "daemon.rb"
       @input = tmp("process_daemon_input_file")
       @data = tmp("process_daemon_data_file")
-      @signal = tmp("process_daemon_signal_file")
       @args = []
     end
 
     def wait_for_daemon
-      10.times do
-        return true if File.exists? @signal and
-                       File.exists? @data and
-                       File.size? @data
-        sleep 0.1
-      end
-
-      return false
+      sleep 0.1 until File.exist?(@data) and File.size?(@data)
     end
 
     def invoke(behavior, arguments=[])
       args = Marshal.dump(arguments).unpack("H*")
-      args << @input << @data << @signal << behavior
+      args << @input << @data << behavior
 
-      ruby_exe @script, :args => args
+      ruby_exe @script, args: args
 
       wait_for_daemon
 
-      return unless File.exists? @data
+      return unless File.exist? @data
 
       File.open(@data, "rb") { |f| return f.read.chomp }
     end
@@ -38,17 +33,23 @@ module ProcessSpecs
   class Signalizer
     attr_reader :pid_file, :pid
 
-    def initialize(scenario=nil, ruby_exe=nil)
+    def initialize(scenario=nil)
+      platform_is :windows do
+        fail "not supported on windows"
+      end
       @script = fixture __FILE__, "kill.rb"
       @pid_file = tmp("process_kill_signal_file")
       rm_r @pid_file
 
       @thread = Thread.new do
-        args = [@pid_file, scenario, ruby_exe]
-        @result = ruby_exe @script, :args => args
+        Thread.current.abort_on_exception = true
+        args = [@pid_file]
+        args << scenario if scenario
+        args << RUBY_EXE.inspect if scenario
+        @result = ruby_exe @script, args: args
       end
-      Thread.pass until File.exists? @pid_file
-      while @pid.nil? || @pid == 0
+      Thread.pass while @thread.status and !File.exist?(@pid_file)
+      while @thread.status && (@pid.nil? || @pid == 0)
         @pid = IO.read(@pid_file).chomp.to_i
       end
     end
@@ -56,7 +57,7 @@ module ProcessSpecs
     def wait_on_result
       # Ensure the process exits
       begin
-        Process.kill :TERM, pid
+        Process.kill :TERM, pid if pid
       rescue Errno::ESRCH
         # Ignore the process not existing
       end
